Add a grain for project domain
In order to ease future migration, I decided to
no longer use foo.gluster.org for any server name, since
this tie the service name to the server. In order to
avoid conflict, machine will include their name, and
the DC they are in in the domain name. So I am gonna
slowly migrate to that in the future.